requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile --resolver=backtracking
#
attrs==22.2.0
    # via pytest
build==0.9.0
    # via pip-tools
click==8.1.3
    # via pip-tools
exceptiongroup==1.0.4
    # via pytest
iniconfig==1.1.1
    # via pytest
mpmath==1.2.1
    # via sympy
numpy==1.24.0
    # via pandas
packaging==22.0
    # via
    #   build
    #   pytest
pandas==1.5.2
    # via src (setup.py)
pep517==0.13.0
    # via build
pip-tools==6.12.1
    # via src (setup.py)
pluggy==1.0.0
    # via pytest
pytest==7.2.0
    # via src (setup.py)
python-dateutil==2.8.2
    # via pandas
pytz==2022.7
    # via pandas
six==1.16.0
    # via python-dateutil
sympy==1.11.1
    # via src (setup.py)
tomli==2.0.1
    # via
    #   build
    #   pep517
    #   pytest
wheel==0.38.4
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
